Implemented tutorial level. Added text to cutscenes and tutorial. Created new "door" prefab for ending levels. Colored bots to reflect behavior. Tweaked some colliders. (#13)
The new "door" prefab is an opening into a room that is filled with an invisible collider, so as soon as the room is entered--as soon as you go through the door--the level transition is triggered. Could be good with a model of doors opening when you get close enough.